sql: move txnState to session.go. No code changes.
It's a member of Session.
Move the planner pool there too; the planner will become a member of the
Session too in the next commit.
